Make the profile page a server component and use a server action to delete a property
