+++ app/profile/page.jsx
@@ -1,76 +1,30 @@
-'use client';
-import { useState, useEffect } from 'react';
 import Image from 'next/image';
-import Link from 'next/link';
-import { useSession } from 'next-auth/react';
 import profileDefault from '@/assets/images/profile.png';
-import Spinner from '@/components/Spinner';
-import { toast } from 'react-toastify';
+import ProfileProperties from '@/components/ProfileProperties';
+import connectDB from '@/config/database';
+import { getSessionUser } from '@/utils/getSessionUser';
+import { convertToSerializeableObject } from '@/utils/convertToObject';
+import Property from '@/models/Property';
 
-const ProfilePage = () => {
-  const { data: session } = useSession();
-  const profileImage = session?.user?.image;
-  const profileName = session?.user?.name;
-  const profileEmail = session?.user?.email;
+// NOTE: This component has been changed to a server component and now fetches
+// the users properties by querying the database directly.
+// We use a new client component - ProfileProperties which takes state from this
+// page component and updates when a property is deleted by using a server
+// action.
 
-  const [properties, setProperties] = useState([]);
-  const [loading, setLoading] = useState(true);
+const ProfilePage = async () => {
+  await connectDB();
 
-  useEffect(() => {
-    const fetchUserProperties = async (userId) => {
-      if (!userId) {
-        return;
-      }
+  const sessionUser = await getSessionUser();
 
-      try {
-        const res = await fetch(`/api/properties/user/${userId}`);
+  const { userId } = sessionUser;
 
-        if (res.status === 200) {
-          const data = await res.json();
-          setProperties(data);
-        }
-      } catch (error) {
-        console.log(error);
-      } finally {
-        setLoading(false);
-      }
-    };
+  if (!userId) {
+    throw new Error('User ID is required');
+  }
 
-    // Fetch user properties when session is available
-    if (session?.user?.id) {
-      fetchUserProperties(session.user.id);
-    }
-  }, [session]);
-
-  const handleDeleteProperty = async (propertyId) => {
-    const confirmed = window.confirm(
-      'Are you sure you want to delete this property?'
-    );
-
-    if (!confirmed) return;
-
-    try {
-      const res = await fetch(`/api/properties/${propertyId}`, {
-        method: 'DELETE',
-      });
-
-      if (res.status === 200) {
-        // Remove the property from state
-        const updatedProperties = properties.filter(
-          (property) => property._id !== propertyId
-        );
-
-        setProperties(updatedProperties);
-
-        toast.success('Property Deleted');
-      } else {
-        toast.error('Failed to delete property');
-      }
-    } catch (error) {
-      console.log(error);
-      toast.error('Failed to delete property');
-    }
-  };
+  const propertiesDocs = await Property.find({ owner: userId }).lean();
+  const properties = propertiesDocs.map(convertToSerializeableObject);
 
   return (
     <section className='bg-blue-50'>
@@ -82,64 +36,28 @@ const ProfilePage = () => {
               <div className='mb-4'>
                 <Image
                   className='h-32 w-32 md:h-48 md:w-48 rounded-full mx-auto md:mx-0'
-                  src={profileImage || profileDefault}
+                  src={sessionUser.user.image || profileDefault}
                   width={200}
                   height={200}
                   alt='User'
                 />
               </div>
               <h2 className='text-2xl mb-4'>
-                <span className='font-bold block'>Name: </span> {profileName}
+                <span className='font-bold block'>Name: </span>{' '}
+                {sessionUser.user.name}
               </h2>
               <h2 className='text-2xl'>
-                <span className='font-bold block'>Email: </span> {profileEmail}
+                <span className='font-bold block'>Email: </span>{' '}
+                {sessionUser.user.email}
               </h2>
             </div>
 
             <div className='md:w-3/4 md:pl-4'>
               <h2 className='text-xl font-semibold mb-4'>Your Listings</h2>
-              {!loading && properties.length === 0 && (
+              {properties.length === 0 ? (
                 <p>You have no property listings</p>
-              )}
-              {loading ? (
-                <Spinner loading={loading} />
               ) : (
-                properties.map((property) => (
-                  <div key={property._id} className='mb-10'>
-                    <Link href={`/properties/${property._id}`}>
-                      <Image
-                        className='h-32 w-full rounded-md object-cover'
-                        src={property.images[0]}
-                        alt=''
-                        width={500}
-                        height={100}
-                        priority={true}
-                      />
-                    </Link>
-                    <div className='mt-2'>
-                      <p className='text-lg font-semibold'>{property.name}</p>
-                      <p className='text-gray-600'>
-                        Address: {property.location.street}{' '}
-                        {property.location.city} {property.location.state}
-                      </p>
-                    </div>
-                    <div className='mt-2'>
-                      <Link
-                        href={`/properties/${property._id}/edit`}
-                        className='bg-blue-500 text-white px-3 py-3 rounded-md mr-2 hover:bg-blue-600'
-                      >
-                        Edit
-                      </Link>
-                      <button
-                        onClick={() => handleDeleteProperty(property._id)}
-                        className='bg-red-500 text-white px-3 py-2 rounded-md hover:bg-red-600'
-                        type='button'
-                      >
-                        Delete
-                      </button>
-                    </div>
-                  </div>
-                ))
+                <ProfileProperties properties={properties} />
               )}
             </div>
           </div>
